Cleand up debug statements, and made the tests look for all called functions in ntdll, passing successfully other than a trace of the CFG into a jump table

+++ Headers/length_disassembly_engine/common.h
@@ -4,5 +4,5 @@
 struct LdeCommon { using enum inst::Context::Status;
     inst::Context         currContext;
     inst::Context::Status status            = success;
-    BYTE                  instruction_count = 0;
+    WORD                  instruction_count = 0;
 };
